+++ cmd/args.go
@@ -8,36 +8,37 @@ import (
 )
 
 //WARNING: the default value field in flag struct is not used
-func setupCommandFlag(settings *rosewood.Settings) (flgSets []*flag.FlagSet, err error) {
+func setupCommandFlag(job *Job) (flgSets []*flag.FlagSet, err error) {
 	globals := NewCommand("", []Flag{
-		{&settings.Debug, "debug", "d", 0},
-		{&settings.MaxConcurrentWorkers, "max-threads", "mt", 0},
+		{&job.Settings.Debug, "debug", "d", 0},
+		{&job.Settings.MaxConcurrentWorkers, "max-threads", "mt", 0},
 	})
+	cmdDo := NewCommand("do", []Flag{})
 	cmdRun := NewCommand("run", []Flag{
-		{&settings.ConvertOldVersions, "convert-old", "co", false},
-		{&settings.ConvertFromVersion, "rosewood-version", "rv", settings.ConvertFromVersion},
-		{&settings.DoNotInlineCSS, "no-inlined-css", "", false},
-		{&settings.OutputFileName, "output", "o", ""},
-		{&settings.OverWriteOutputFile, "replace", "r", false},
-		{&settings.PreserveWorkFiles, "keep-temp", "k", false},
-		{&settings.SaveConvertedFile, "save-converted", "sc", false},
-		//		{&settings.SectionSeparator, "sep", "S", "+++"},
-		{&settings.StyleSheetName, "style", "s", ""},
-		{&settings.WorkDirName, "work-dir", "w", ""},
+		{&job.Settings.ConvertOldVersions, "convert-old", "co", false},
+		{&job.Settings.ConvertFromVersion, "rosewood-version", "rv", job.Settings.ConvertFromVersion},
+		{&job.Settings.DoNotInlineCSS, "no-inlined-css", "", false},
+		{&job.OutputFile.Name, "output", "o", ""},
+		{&job.Settings.OverWriteOutputFile, "replace", "r", false},
+		{&job.Settings.PreserveWorkFiles, "keep-temp", "k", false},
+		{&job.Settings.SaveConvertedFile, "save-converted", "sc", false},
+		//	S{&job.Settings.SectionSeparator, "sep", "S", "+++"},
+		{&job.Settings.StyleSheetName, "style", "s", ""},
+		{&job.WorkDirName, "work-dir", "w", ""},
 	})
 	cmdCheck := NewCommand("check", []Flag{
-		//		{&settings.SectionSeparator, "sep", "S", "+++"},
+		//		{&job.Settings.SectionSeparator, "sep", "S", "+++"},
 	})
 	cmdV1tov2 := NewCommand("v1tov2", []Flag{
-		{&settings.ConvertFromVersion, "rosewood-version", "rv", ""},
-		{&settings.OverWriteOutputFile, "replace", "r", false},
+		{&job.Settings.ConvertFromVersion, "rosewood-version", "rv", ""},
+		{&job.Settings.OverWriteOutputFile, "replace", "r", false},
 	})
 	cmdInit := NewCommand("init", []Flag{
-		{&settings.OverWriteOutputFile, "replace", "r", false},
+		{&job.Settings.OverWriteOutputFile, "replace", "r", false},
 	})
 	cmdHelp := NewCommand("help", []Flag{})
 	cmdVersion := NewCommand("version", []Flag{})
-	flgSets = append(flgSets, globals, cmdRun, cmdCheck, cmdV1tov2, cmdHelp, cmdVersion, cmdInit)
+	flgSets = append(flgSets, globals, cmdDo, cmdRun, cmdCheck, cmdV1tov2, cmdHelp, cmdVersion, cmdInit)
 	for _, fs := range flgSets {
 		fs.Usage = func() {}    //disable internal usage function
 		fs.SetOutput(devNull{}) //suppress output from package flag