+++ cmd/carpenter.go
@@ -4,8 +4,9 @@
 package main
 
 import (
+	"encoding/json"
 	"fmt"
-	"io"
+	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
@@ -16,12 +17,6 @@ import (
 	rosewood "github.com/drgo/rosewood/lib"
 	_ "github.com/drgo/rosewood/renderers/html" //include needed renderers
 )
-
-type result struct {
-	inputFileName  string
-	outputFileName string
-	err            error
-}
 
 const (
 	ConfigFileBaseName = "carpenter.json"
@@ -40,7 +35,6 @@ const (
 // add graceful shutdown https://golang.org/pkg/os/signal/
 // add cleanup routine
 // add reading of settings from a json file specified using config command--see setters.go
-// add support for running jobs from a json file
 // add support for settings in package types
 // clean-up all tests.
 // add css into generated html files
@@ -60,47 +54,44 @@ func main() {
 //TODO: add gracefull shutdown logic here along with a pointer to an optional cleanup function
 
 //RunApp has all program logic; entry point for all tests
+//WARNING: not safe to call concurrently; this is the only function allowed to change the job configuration
+//WARNING: beyond this function, changing settings or job fields is not permitted
 func RunApp() error {
-	//fmt.Printf("main: len=%d ==>%v\n", len(cmdArgs), cmdArgs)
-	if len(os.Args) == 1 { //only app name passed, exit
-		return fmt.Errorf(ErrWrongCommand)
+	if len(os.Args) == 1 { //no command line arguments
+		return DoFromConfigFile()
 	}
-	//load configuration from config file if one exists in current dir
-	settings, err := GetValidSettings()
+	job, err := LoadConfigFromCommandLine()
 	if err != nil {
 		return err
 	}
-	settings.ExecutableVersion = Version
-	flgSets, _ := setupCommandFlag(settings)
-	flg, err := ParseCommandLine(flgSets[0], flgSets[1:]...)
-	if err != nil {
-		return err
+	if job.Settings.Debug == rosewood.DebugAll {
+		fmt.Printf("current settings:\n%s\n", job)
 	}
-	if settings.Debug == rosewood.DebugAll {
-		fmt.Printf("current settings:\n%s\n", settings)
-	}
-	inputFileNames := flg.Args()
-	switch flg.Name() {
+	switch job.Command { //TODO: check command is case insensitive
+	case "do":
+		if err = DoFromConfigFile(); err != nil {
+			return err
+		}
 	case "check":
-		settings.CheckSyntaxOnly = true
+		job.Settings.CheckSyntaxOnly = true //TODO: rationalize; not needed any more here
 		fallthrough
 	case "run":
 		//FIXME: this check is not working
-		if err = Run(settings, inputFileNames); rosewood.Errors().IsParsingError(err) {
+		if err = DoRun(job); rosewood.Errors().IsParsingError(err) {
 			err = fmt.Errorf("one or more errors occurred during file processing") //do not report again
 		}
 	case "v1tov2":
-		if err = V1toV2(settings, inputFileNames); err != nil { //shadwing err, so check/run/convert errors will not returned
+		if err = V1toV2(job); err != nil { //shadwing err, so check/run/convert errors will not returned
 			fmt.Printf("one or more errors occurred during file processing")
 		}
-	case "init":
-		if err = Init(settings, inputFileNames); err != nil {
+	case "init": //FIXME
+		if err = DoInit(job); err != nil {
 			fmt.Printf("one or more errors occurred during configuration initialization")
 		}
 	case "version":
 		fmt.Println(getVersion())
-	case "help":
-		helpMessage(flg.Args(), getVersion())
+	case "help": //FIXME:
+		//helpMessage(job.InputFiles, getVersion())
 	default:
 		helpMessage(nil, getVersion())
 		return fmt.Errorf(ErrWrongCommand)
@@ -108,193 +99,131 @@ func RunApp() error {
 	return err
 }
 
-//Run is the main work-horse function;
-func Run(settings *rosewood.Settings, inputFileNames []string) error {
-	//No arguments: check if the input is coming from stdin
-	if len(inputFileNames) == 0 {
-		if info, _ := os.Stdin.Stat(); info.Size() == 0 {
-			return fmt.Errorf(ErrMissingInFile)
-		}
-		inputFileNames = append(inputFileNames, "") //empty argument signals stdin
-	}
+func DoFromConfigFile() error {
 	var (
+		configFileName string
 		err            error
-		baseDir        string
-		format         string
-		outputFileName = strings.TrimSpace(settings.OutputFileName)
-		start          time.Time
 	)
-	verb := "Check"
-	if !settings.CheckSyntaxOnly {
-		if format, err = GetValidFormat(inputFileNames, outputFileName); err != nil {
+	if len(os.Args) == 1 { //only app name passed, use ConfigFileBaseName in current folder
+		if configFileName, err = fileutils.GetFullPath(ConfigFileBaseName); err != nil {
 			return err
 		}
-		preserveWorkFiles := strings.TrimSpace(settings.WorkDirName) != "" || (format == "docx" && settings.PreserveWorkFiles) || format == "html"
-		if baseDir, err = GetOutputBaseDir(settings.WorkDirName, preserveWorkFiles); err != nil {
+	} else {
+		//2 or more arguments, is it app name + do + a json file
+		if strings.TrimSpace(strings.ToLower(os.Args[1])) != "do" {
+			return fmt.Errorf("invalid command %s", os.Args[1])
+		}
+		if len(os.Args) < 3 {
+			return fmt.Errorf("must specify a json configuration file")
+		}
+		configFileName = os.Args[2]
+		if strings.ToLower(filepath.Ext(configFileName)) != ".json" {
+			return fmt.Errorf("invalid config file name [%s] passed, ext must be json", configFileName)
+		}
+		if configFileName, err = fileutils.GetFullPath(configFileName); err != nil {
 			return err
 		}
-		if !preserveWorkFiles { //baseDir is temp, schedule removing it
-			defer os.RemoveAll(baseDir)
-		}
-		verb = "Process"
 	}
-
-	if settings.Debug >= rosewood.DebugUpdates {
-		fmt.Printf("%sing %d file(s)\n", verb, len(inputFileNames))
-		start = time.Now()
+	//load configuration from config file
+	configBuf, err := ioutil.ReadFile(configFileName)
+	if err != nil {
+		return fmt.Errorf("failed to load configuration file: %v", err)
 	}
-
-	processedFiles, err := runHTMLFiles(settings, inputFileNames, baseDir, format)
-	if err != nil || len(processedFiles) == 0 {
-		//	fmt.Printf("runHTMLFiles failed %s", err)
-		return err
+	job := DefaultJob(rosewood.DefaultSettings())
+	if err = json.Unmarshal(configBuf, job); err != nil {
+		return fmt.Errorf("failed to parse configuration file %s: %v", configFileName, err)
 	}
-	//fmt.Printf("format=%s \n %s \n", format, strings.Join(processedFiles, "|"))
-	if !settings.CheckSyntaxOnly {
-		switch {
-		case format == "docx":
-			docxOpts := htmldocx.DefaultOptions().SetDebug(settings.Debug)
-			if err = htmldocx.Convert(processedFiles, outputFileName, docxOpts); err != nil {
-				return fmt.Errorf("failed to convert to docx file: %s", err)
-			}
-		case format == "html":
-			// fmt.Println("about to save file as " + outputFileName)
-			// if outputFileName != "" { //we are here, so there must be only 1 file in processedFiles, rename it
-			// 	if err = os.Rename(processedFiles[0], outputFileName); err != nil {
-			// 		return fmt.Errorf("failed to rename file to %s: %s", outputFileName, err)
-			// 	} else {
-			// 		fmt.Println("file saved as " + outputFileName)
-			// 	}
-			// }
-		default:
-			return fmt.Errorf("unsupported format: %s", format) //should not happen
-		}
+	if job.Settings.Debug >= rosewood.DebugUpdates {
+		fmt.Println("configuration loaded from " + configFileName)
 	}
-	if settings.Debug >= rosewood.DebugUpdates {
-		fmt.Printf("%sed %d file(s) in %s\n", verb, len(inputFileNames), time.Since(start).String())
+	job.FileName = configFileName
+	if job.Settings.Debug >= rosewood.DebugAll {
+		fmt.Printf("current configuration: \n %s\n", job)
+	}
+	if err = DoRun(job); rosewood.Errors().IsParsingError(err) {
+		err = fmt.Errorf("one or more errors occurred during file processing") //do not report again
 	}
 	return err
 }
 
-//basedir always points to a valid dir to save output files
-func runHTMLFiles(settings *rosewood.Settings, inputFileNames []string, baseDir, format string) ([]string, error) {
-	report := func(res result) {
-		if !settings.CheckSyntaxOnly {
-			fmt.Printf("\n--------------\nprocessing %s:", res.inputFileName)
-		} else {
-			fmt.Printf("\n--------------\nchecking %s:", res.inputFileName)
-		}
-		if res.err != nil {
-			fmt.Printf("\nErrors: %v\n", res.err)
-		} else {
-			fmt.Printf("...Done\n")
-			if !settings.CheckSyntaxOnly {
-				fmt.Printf("output file: %s\n", res.outputFileName)
-			}
-		}
-	}
-	//channel to communicate with
-	resCh := make(chan result)
-	//A counting semaphore to limit number of open files
-	tokens := NewCountingSemaphore(settings.MaxConcurrentWorkers)
-	go func() {
-		outputFileName := strings.TrimSpace(settings.OutputFileName)
-		tempOutputFileName := ""
-		for _, inputFileName := range inputFileNames {
-			switch {
-			case outputFileName == "": //no output file assume html file with the same base name as inputfile
-				tempOutputFileName = filepath.Join(baseDir,
-					fileutils.ReplaceFileExt(filepath.Base(inputFileName), "html"))
-			case format == "html": //this happens only if there was a single inputfile
-				//AND outfilename with html ext, use the outputfilename
-				if filepath.Dir(outputFileName) == "." { //no directory, use the baseDir
-					tempOutputFileName = filepath.Join(baseDir, outputFileName)
-				}
-			case format == "docx": //create temp html files in the basedir
-				tempOutputFileName = filepath.Join(baseDir,
-					fileutils.ReplaceFileExt(filepath.Base(inputFileName), "html"))
-			default:
-				panic("unexpected branch in runHTMLFiles()") //should not happen
-			}
-			fmt.Println(outputFileName + "-->" + tempOutputFileName)
-			tokens.Reserve(1)                                                 //reserve a worker
-			go htmlRunner(settings, inputFileName, tempOutputFileName, resCh) //launch a runSingle worker for each file
-		}
-	}()
-	var err error
-	var processedFiles []string
-	for i := 0; i < len(inputFileNames); i++ { //wait for workers to return one by one
-		//fmt.Println("inside for loop")
-		res := <-resCh
-		//fmt.Printf("%+v", res)
-		tokens.Free(1) //release a reserved worker
-		if settings.Debug >= rosewood.DebugUpdates || res.err != nil {
-			report(res)
-		}
-		if res.err == nil {
-			processedFiles = append(processedFiles, res.outputFileName)
-		}
-		if err == nil {
-			err = res.err
-		}
-	}
-	return processedFiles, err
-}
-
-//htmlRunner parses and renders (if in run mode) a single input file into an HTML file
-//all errors are returned through resChan channel; only one error per run
-func htmlRunner(settings *rosewood.Settings, inputFileName, outputFileName string, resChan chan result) {
-	var (
-		in  *os.File
-		out *os.File
-		err error //not returned, used to decide whether the output file should be saved or not
-	)
-	if inputFileName == "" { //reading from stdin
-		inputFileName = "<stdin>"
-	}
-	iDesc := DefaultRwInputDescriptor(settings)
-	if in, err = getValidInputReader(iDesc.SetFileName(inputFileName)); err != nil {
-		resChan <- result{inputFileName, outputFileName, err}
-		return
-	}
-	defer in.Close()
-	if !settings.CheckSyntaxOnly { //do not need an output
-		//if the outputFileName already exists and OverWriteOutputFile is false, return an error
-		if _, err := os.Stat(outputFileName); err == nil && !settings.OverWriteOutputFile {
-			resChan <- result{inputFileName, outputFileName, fmt.Errorf("file already exists: %s", outputFileName)}
-			return
-		}
-		//output either settings.OutputFileName or a new file =inFileName + "ext" if settings.OutputFileName==""
-		if out, err = getOutputWriter(outputFileName, settings.OverWriteOutputFile); err != nil {
-			resChan <- result{inputFileName, outputFileName, err}
-			return
-		}
-		//define a function that saves the temp output file created below using settings.OutputFileName
-		defer func() {
-			if err == nil { //only save temp file if runFile() below succeeded
-				resChan <- result{inputFileName, outputFileName, fileutils.CloseAndRename(out, outputFileName, settings.OverWriteOutputFile)}
-				return
-			}
-		}()
-	}
-	ri := rosewood.NewInterpreter(settings).SetScriptIdentifer(inputFileName)
-	err = runFile(ri, in, out)
-	resChan <- result{inputFileName, outputFileName, err}
-}
-
-func runFile(ri *rosewood.Interpreter, in io.ReadSeeker, out io.Writer) error {
-	file, err := ri.Parse(in, ri.ScriptIdentifer())
-	if err != nil || ri.Setting().CheckSyntaxOnly {
-		return ri.ReportError(err)
-	}
-	hr, err := rosewood.GetRendererByName("html") //TODO: get from settings
+func LoadConfigFromCommandLine() (*Job, error) {
+	job := DefaultJob(rosewood.DefaultSettings()) //TODO: ensure all defaults are reasonable
+	flgSets, _ := setupCommandFlag(job)
+	flg, err := ParseCommandLine(flgSets[0], flgSets[1:]...)
 	if err != nil {
+		return nil, err
+	}
+	job.Command = flg.Name()
+	//TODO: validate command line inputs; use in RunfromConfigFile too?!
+	for _, fileName := range flg.Args() {
+		job.InputFiles = append(job.InputFiles, NewFileDescriptor(fileName))
+	}
+	return job, nil
+}
+
+//Run is the main work-horse function;
+//WARNING: not safe to call concurrently
+func DoRun(job *Job) error {
+	var (
+		err   error
+		start time.Time
+	)
+	//if no input files: check if the input is coming from stdin
+	if len(job.InputFiles) == 0 {
+		if info, _ := os.Stdin.Stat(); info.Size() == 0 {
+			return fmt.Errorf(ErrMissingInFile)
+		}
+		job.InputFiles = append(job.InputFiles, NewFileDescriptor("")) //empty argument signals stdin
+	}
+	if !job.Settings.CheckSyntaxOnly {
+		if job.OutputFormat, err = GetValidFormat(job); err != nil {
+			return err
+		}
+		job.WorkDirName = strings.TrimSpace(job.WorkDirName)
+		fmt.Printf("job.WorkDirName =%v \n", job.WorkDirName)
+		job.PreserveWorkFiles = job.OutputFormat == "html" || job.WorkDirName != "" || (job.OutputFormat == "docx" && job.Settings.PreserveWorkFiles)
+		fmt.Printf("job.PreserveWorkFiles=%v, %v, %v \n", job.OutputFormat == "html", job.WorkDirName != "", (job.OutputFormat == "docx" && job.Settings.PreserveWorkFiles))
+		if job.WorkDirName, err = GetOutputBaseDir(job.WorkDirName, job.PreserveWorkFiles); err != nil {
+			return err
+		}
+		if !job.PreserveWorkFiles { //baseDir is temp, schedule removing it
+			defer os.RemoveAll(job.WorkDirName)
+		}
+	}
+
+	if job.Settings.Debug >= rosewood.DebugUpdates {
+		fmt.Printf("%sing %d file(s) in work dir=%s\n", job.Command, len(job.InputFiles), job.WorkDirName)
+		start = time.Now()
+	}
+
+	processedFiles, err := runHTMLFiles(job)
+	if err != nil || len(processedFiles) == 0 {
 		return err
 	}
-	return ri.ReportError(ri.Render(out, file, hr))
+	fmt.Printf("format=%s \n %s \n", job.OutputFormat, strings.Join(processedFiles, "|\n"))
+	if !job.Settings.CheckSyntaxOnly {
+		switch {
+		case job.OutputFormat == "docx":
+			docxOpts := htmldocx.DefaultOptions().SetDebug(job.Settings.Debug)
+			if err = htmldocx.Convert(processedFiles, job.OutputFile.Name, docxOpts); err != nil {
+				return fmt.Errorf("failed to convert to docx file: %s", err)
+			}
+			if job.Settings.Debug >= rosewood.DebugUpdates {
+				fmt.Printf("saved to docx %s\n", job.OutputFile.Name)
+			}
+
+		case job.OutputFormat == "html":
+		default:
+			return fmt.Errorf("unsupported format: %s", job.OutputFormat) //should not happen
+		}
+	}
+	if job.Settings.Debug >= rosewood.DebugUpdates {
+		fmt.Printf("%sed %d file(s) in %s\n", job.Command, len(job.InputFiles), time.Since(start).String())
+	}
+	return err
 }
 
-// func runMulti(settings *rosewood.Settings, inFileNames []string, resChan chan result) {
+// func runMulti(settings *rosewood.Settings, inFileNames []string, resChan chan task) {
 // 	var (
 // 		in  *os.File
 // 		out *os.File
@@ -304,13 +233,13 @@ func runFile(ri *rosewood.Interpreter, in io.ReadSeeker, out io.Writer) error {
 // 	//define a function that saves the temp output file created below using settings.OutputFileName
 // 	onOutputFileClose := func() {
 // 		if err == nil { //only save temp file if runFile() below succeeded
-// 			resChan <- result{outputFileName, fileutils.CloseAndRename(out, outputFileName, settings.OverWriteOutputFile)}
+// 			resChan <- task{outputFileName, fileutils.CloseAndRename(out, outputFileName, settings.OverWriteOutputFile)}
 // 			return
 // 		}
 // 	}
 // 	//open output file
 // 	if out, err = getOutputFile(outputFileName, settings.OverWriteOutputFile); err != nil {
-// 		resChan <- result{outputFileName, err}
+// 		resChan <- task{outputFileName, err}
 // 		return
 // 	}
 // 	defer onOutputFileClose()
@@ -319,12 +248,12 @@ func runFile(ri *rosewood.Interpreter, in io.ReadSeeker, out io.Writer) error {
 // 	for _, inputFileName := range inFileNames {
 // 		in, err = getValidInputReader(DefaultRwInputDescriptor(settings).SetFileName(inputFileName))
 // 		if err != nil {
-// 			resChan <- result{inputFileName, err}
+// 			resChan <- task{inputFileName, err}
 // 			return
 // 		}
 // 		defer in.Close()
 // 		if err = runFile(ri, in, out); err != nil {
-// 			resChan <- result{inputFileName, err}
+// 			resChan <- task{inputFileName, err}
 // 			return
 // 			//errs.Add(fmt.Errorf("error running file %s:\n%s", inputFileName, errors.ErrorsToError(err)))
 // 		}
@@ -363,7 +292,7 @@ func runFile(ri *rosewood.Interpreter, in io.ReadSeeker, out io.Writer) error {
 
 // if settings.OutputFileName != "" { // && len(args) > 1
 //multiple files with one single output file
-// if settings.Debug >= rosewood.DebugUpdates {
+// if job.Settings.Debug >= rosewood.DebugUpdates {
 // 	fmt.Println("in a multi-input single-output mode")
 // }
 // if err = runMulti(settings, args); err != nil {
@@ -372,6 +301,6 @@ func runFile(ri *rosewood.Interpreter, in io.ReadSeeker, out io.Writer) error {
 // } else {
 //either one input file with
 //this signals that we need to create one outputfile for each input file
-// if settings.Debug >= rosewood.DebugUpdates {
+// if job.Settings.Debug >= rosewood.DebugUpdates {
 // 	fmt.Println("in a multi-input multi-output mode")
 // }