+++ cmd/init.go
@@ -4,22 +4,18 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-
-	"github.com/drgo/rosewood/lib"
-	"github.com/drgo/rosewood/lib/setter"
 )
 
-func Init(settings *rosewood.Settings, args []string) error {
-	settings = rosewood.DefaultSettings()
+func DoInit(job *Job) error {
 	var path string
-	if len(args) == 0 {
+	if job.GetNameOfInputFile(0) == "" { //first input file is the name of new config file
 		dir, err := os.Getwd()
 		if err != nil {
 			return fmt.Errorf("failed to create config file in dir %s: %s", dir, err)
 		}
 		path = filepath.Join(dir, ConfigFileBaseName)
 	} else {
-		path = args[0]
+		path = job.InputFiles[0].Name
 	}
-	return setter.SaveSettings(settings, path, true)
+	return SaveJob(job, path, true)
 }